+++ src/pages/api/hello.ts
@@ -5,7 +5,7 @@ type Data = {
   name: string;
 };
 
-const handler = (req: NextApiRequest, res: NextApiResponse<Data>) => {
+const handler = (_req: NextApiRequest, res: NextApiResponse<Data>) => {
   res.status(200).json({ name: 'John Doe' });
 };